bugfix - grafana values template has a misplaced end tag
the end tag's current placement means that if monitoring
is turned off and istio is turned on, then a block containing
istio data meant to be used by monitoring will be rendered
but the accompanying block within which that data should be
enclosed is not rendered.  this leads to superflous data
being added to the grafana.ini block in some cases and
in other cases, like when sso is turned on and configured,
malformed yaml.